Add `cargo zng l10n --translate`
