+++ crates/cargo-zng/src/l10n.rs
@@ -17,29 +17,32 @@ use clap::*;
 
 use crate::{l10n::scraper::FluentTemplate, util};
 
-mod pseudo;
 mod scraper;
+
+mod generate_util;
+mod pseudo;
+mod translate;
 
 #[derive(Args, Debug)]
 pub struct L10nArgs {
     /// Rust files glob or directory
-    #[arg(short, long, default_value = "")]
+    #[arg(short, long, default_value = "", value_name = "PATH", hide_default_value = true)]
     input: String,
 
     /// L10n resources dir
-    #[arg(short, long, default_value = "")]
+    #[arg(short, long, default_value = "", value_name = "DIR", hide_default_value = true)]
     output: String,
 
     /// Package to scrap and copy dependencies
     ///
     /// If set the --input and --output default is src/**.rs and l10n/
-    #[arg(short, long, default_value = "")]
+    #[arg(short, long, default_value = "", hide_default_value = true)]
     package: String,
 
     /// Path to Cargo.toml of crate to scrap and copy dependencies
     ///
     /// If set the --input and --output default to src/**.rs and l10n/
-    #[arg(long, default_value = "")]
+    #[arg(long, default_value = "", hide_default_value = true)]
     manifest_path: String,
 
     /// Don't copy dependencies localization
@@ -73,7 +76,7 @@ pub struct L10nArgs {
     clean: bool,
 
     /// Custom l10n macro names, comma separated
-    #[arg(short, long, default_value = "")]
+    #[arg(short, long, default_value = "", hide_default_value = true)]
     macros: String,
 
     /// Generate pseudo locale from dir/lang
@@ -81,14 +84,47 @@ pub struct L10nArgs {
     /// EXAMPLE
     ///
     /// "l10n/en" generates pseudo from "l10n/en.ftl" and "l10n/en/*.ftl"
-    #[arg(long, default_value = "")]
+    #[arg(long, default_value = "", value_name = "PATH", hide_default_value = true)]
     pseudo: String,
     /// Generate pseudo mirrored locale
-    #[arg(long, default_value = "")]
+    #[arg(long, default_value = "", value_name = "PATH", hide_default_value = true)]
     pseudo_m: String,
     /// Generate pseudo wide locale
-    #[arg(long, default_value = "")]
+    #[arg(long, default_value = "", value_name = "PATH", hide_default_value = true)]
     pseudo_w: String,
+
+    /// Generate machine translation from dir/lang
+    ///
+    /// SERVICE
+    ///
+    /// An external translation service must be configured in an env var
+    ///
+    /// CARGO_ZNG_TRANSLATE=service --from {from} --to {to} -i {text} --context-info {comments}
+    ///
+    /// The {from}, {to} and {text} are required
+    ///
+    /// The translated text is read from stdout
+    ///
+    /// See github.com/zng-ui/zng/blob/main/docs/cargo-zng-translate.md for mode details
+    ///
+    /// EXAMPLE
+    ///
+    /// "l10n/template" generates translation from "l10n/template.ftl" and "l10n/template/*.ftl" for each
+    /// --translate-to language
+    #[arg(long, default_value = "", value_name = "PATH", hide_default_value = true)]
+    translate: String,
+    /// Target locales for --translate
+    ///
+    /// # EXAMPLE
+    ///
+    /// "es,pt" translates to spanish and portuguese
+    #[arg(long, value_name = "LANGS", value_delimiter = ',')]
+    translate_to: Vec<String>,
+    /// Source locale for --translate
+    ///
+    /// Default is the folder name, or "en" for "template"
+    #[arg(long, default_value = "", hide_default_value = true)]
+    translate_from: String,
 
     /// Verify that packages are scrapped and validate Fluent files
     #[arg(long, action)]
@@ -164,7 +200,7 @@ pub fn run(mut args: L10nArgs) {
     }
 
     if input.is_empty() {
-        return run_pseudo(&args);
+        return run_generators(&args);
     }
 
     if output.is_empty() {
@@ -234,7 +270,7 @@ pub fn run(mut args: L10nArgs) {
         check_fluent_output(&args, output);
     }
 
-    run_pseudo(&args);
+    run_generators(&args);
 }
 
 fn check_scrap_package(args: &L10nArgs, input: &str, output: &Path, template: &mut FluentTemplate) {
@@ -461,6 +497,9 @@ fn check_scrap_package(args: &L10nArgs, input: &str, output: &Path, template: &m
                     check: args.check,
                     check_strict: args.check_strict,
                     verbose: args.verbose,
+                    translate: String::new(),
+                    translate_from: String::new(),
+                    translate_to: vec![],
                 },
                 &input,
                 output,
@@ -470,7 +509,10 @@ fn check_scrap_package(args: &L10nArgs, input: &str, output: &Path, template: &m
     }
 }
 
-fn run_pseudo(args: &L10nArgs) {
+fn run_generators(args: &L10nArgs) {
+    if !args.translate.is_empty() {
+        translate::translate(&args.translate, &args.translate_from, &args.translate_to, args.check, args.verbose);
+    }
     if !args.pseudo.is_empty() {
         pseudo::pseudo(&args.pseudo, args.check, args.verbose);
     }
